refactor(Label.stories): rename the files by prefixing them with the component name
