+++ src/components/inputs/Label/stories/stories.args.js
